adjusted db to hold double values

+++ EtcsServer/Database/Entity/RailroadSign.cs
@@ -5,8 +5,8 @@
         public int RailroadSignId { get; set; }
         public int TrackId { get; set; }
         public Track Track { get; set; }
-        public int DistanceFromTrackStart { get; set; }
+        public double DistanceFromTrackStart { get; set; }
         public bool IsFacedUp { get; set; }
-        public int MaxSpeed { get; set; }
+        public double MaxSpeed { get; set; }
     }
 }
